+++ backend2/src/main/java/com/cmoney/backend2/brokerdatatransmission/service/api/brokerstockdata/put/StockInfo.kt
@@ -1,0 +1,13 @@
+package com.cmoney.backend2.brokerdatatransmission.service.api.brokerstockdata.put
+
+import com.cmoney.backend2.brokerdatatransmission.service.api.brokerstockdata.TradeType
+import com.google.gson.annotations.SerializedName
+
+class StockInfo(
+    @SerializedName("TradeType")
+    val tradeType: TradeType,
+    @SerializedName("Amount")
+    val amount: Int,
+    @SerializedName("TradeTotalCost")
+    val tradeTotalCost: Double
+)