+++ backend2/src/main/java/com/cmoney/backend2/brokerdatatransmission/service/api/encryptionkey/GetEncryptionKeyResponse.kt
@@ -1,0 +1,5 @@
+package com.cmoney.backend2.brokerdatatransmission.service.api.encryptionkey
+
+data class GetEncryptionKeyResponse(
+    val publicKeyCryptography: String?
+)